+++ pkg/apis/kubernetes/v1/cluster_types.go
@@ -1,6 +1,7 @@
 package v1
 
 import (
+	"go.anx.io/go-anxcloud/pkg/apis/common"
 	"go.anx.io/go-anxcloud/pkg/apis/common/gs"
 	corev1 "go.anx.io/go-anxcloud/pkg/apis/core/v1"
 )
@@ -30,6 +31,29 @@ type Cluster struct {
 	// If enabled, Service VMs are set up as LBaaS hosts enabling K8s services of type LoadBalancer.
 	// Requires Service VMs.
 	EnableLBaaS *bool `json:"enable_lbaas,omitempty"`
+
+	// Identifier of an internal v4 prefix (to be) assigned to the cluster. If ManageInternalIPv4Prefix
+	// is set to false, the Prefix given in this field is used when creating the cluster, otherwise a new
+	// prefix will be created automatically. The API will always return the Prefix for the Cluster,
+	// when ManageInternalIPv4Prefix is true, this will be the Prefix that was created automatically.
+	InternalIPv4Prefix *common.PartialResource `json:"internal_ipv4_prefix,omitempty"`
+	// Identifier of an external v4 prefix (to be) assigned to the cluster. If ManageExternalIPv4Prefix
+	// is set to false, the Prefix given in this field is used when creating the cluster, otherwise a new
+	// prefix will be created automatically. The API will always return the Prefix for the Cluster,
+	// when ManageExternalIPv4Prefix is true, this will be the Prefix that was created automatically.
+	ExternalIPv4Prefix *common.PartialResource `json:"external_ipv4_prefix,omitempty"`
+	// Identifier of an external v6 prefix (to be) assigned to the cluster. If ManageExternalIPv6Prefix
+	// is set to false, the Prefix given in this field is used when creating the cluster, otherwise a new
+	// prefix will be created automatically. The API will always return the Prefix for the Cluster,
+	// when ManageExternalIPv6Prefix is true, this will be the Prefix that was created automatically.
+	ExternalIPv6Prefix *common.PartialResource `json:"external_ipv6_prefix,omitempty"`
+
+	// If set to true an internal v4 prefix is automatically created for the cluster. Defaults to true if not set.
+	ManageInternalIPv4Prefix *bool `json:"manage_internal_ipv4_prefix,omitempty"`
+	// If set to true an external v4 prefix is automatically created for the cluster. Defaults to true if not set.
+	ManageExternalIPv4Prefix *bool `json:"manage_external_ipv4_prefix,omitempty"`
+	// If set to true an external v6 prefix is automatically created for the cluster. Defaults to true if not set.
+	ManageExternalIPv6Prefix *bool `json:"manage_external_ipv6_prefix,omitempty"`
 
 	// Contains a kubeconfig if available
 	KubeConfig *string `json:"kubeconfig,omitempty"`