+++ CHANGELOG.md
@@ -29,6 +29,10 @@ Some examples, more below in the actual changelog (newer entries are more likely
 ### Fixed
 * trace logging in pkg/client now really includes the request/response bodies (#211, @LittleFox94)
 
+### Added
+* kubernetes/v1: configurable cluster prefixes (#208, @marioreggiori)
+* generic client: common resource types (#208, @marioreggiori)
+
 ## [0.5.0] - 2022-11-30
 
 ### Added